Adding design data to editor (#62)
Sets up fake data for the XAML designer so that it is easier to see what
the editor window will look like when working with it. Should have no
impact on the runtime functionality as it just provides a data context
for the designer.

- EditorShellControl.xaml now uses new EditorShellDesignViewModel
instead of IEditorShellViewModel
- RubberduckEditorControl.xaml now uses CodePaneDesignViewModel instead
of ICodePaneViewModel
- StatusBarControl.xaml now uses StatusBarDesignViewModel instead of
IStatusBarViewModel
- SyncPanelToolControl.xaml now uses SyncPanelDesignViewModel instead of
ISyncPanelToolTab

Each of the 4 controls have data for viewing and all come through to
EditorShellControl.xaml to look like this in VS:

<img width="803" alt="EditorDesignView"
src="https://user-images.githubusercontent.com/4892984/233693369-8dae4eab-2638-4a9c-8223-da904a2d5739.png">

I don't know if this is worth having at this stage in the project but I
found it a nice learning exercise anyway.