Fix script-scope variable declaration bug (#64)
Fixes a bug that presents as `Get-ItemPropertyValue` error messages when
converting from RD3 to RD2.

Despite the error messages, the conversion to RD2 succeeds. However,
going back from RD2 to RD3 will fail.
The user then has to build the Rubberduck3 solution in order to setup
the registry for RD3.

The change addresses the error messages and enables the ability to
toggle versions without re-building.

Also includes the removal of an un-used function (un-related to the
bug).